Fix #395: get warp working on f36
The latest version of warp compiles but when run on the Sirepo EGun
example it ends in a segfault. So, we went back to the version of warp
we were using on f32. Somehow the commit we were using in f32 is no
longer python 3 compatible. So, run lib2to3 and fixup the install to
make it work in our current environment.