Fix #350: Fix nfs mounts for fedora 36.
The root cause of the problems is that vagrant doesn't support
NetworkManager which is used in Fedora 36
https://github.com/hashicorp/vagrant/issues/12762. So, the
private_network we define in the Vagrantfile gets setup on the host
but not in the guest. We fix this by using vagrant_dev_provision_eth1
(set the interface ourselves). But, this provisioner runs after the
sync_folder(s) are created. So, the synced_folders fail when using nfs
because networking isn't setup right.

There is no way to run a provisioner before a synced_folder,
https://github.com/hashicorp/vagrant/issues/936. So, we must set
`disabled: true`, bring the vm up, set `disabled: false`, and reload
the vm.

In addition, git introduced checking of owners for directories
https://github.com/git/git/commit/8959555cee7ec045958f9b6dd62e541affb7e7d9.
In the NFS case the UID/GID from the host likely doesn't match the
UID/GID of the user in the vm. This causes git to complain. So, we
need to set safe.directory to '*'. This configuration is set inside of
the vm and we only have one user inside of vms so it is safe.