Fix #467: fix mantid on f36
Remove opencascad-devel which brought in 3GB of dependencies. It caused
the error on import, was large, and was unused because we don't
do any rendering with Mantid. In addition, disable other unused parts
of Mantid (ex docs).

Add ${module}_test which we can use for adding tests to modules which
will hopefully prevent regressions like this from happening in the future.