Fix #510: remove numpy pin (#511)
In https://github.com/numpy/numpy/releases/tag/v1.25.2 numpy fixed the
issuse that causes `from numpy import *` to override builtins like min
and max.

They've promised to keep this behavior until numpy 2.0 when `from numpy import *`
will once again override the builtins. I chose not to put a pin to <2.0
so we know when 2.0 is released and we can test against it. If 2.0 proves
to be too big of a change we can pin then.