Fix #590: update tensorflow to 2.15.0
- update tensorflow to 2.15.0
- xraylib from rscode-xraylib conflicted with rscode-bluesky so it was moved into common.
After this move there was nothing left in xraylib.sh so it was removed
- Removed some feature tests for f36 vs f32. There are tests in other places but I removed
ones directly related to my work. We are going to be fully on f36 so all tests can be removed.
- update h5py to 3.10.0. This wasn't strictly necessary. Tensorflow can use and older h5py
but since I was in here updating tensorflow I figured it was good to update related packages.