Fix #594: create installer for beamsim jupyter scientific packages
Dependencies like jupyterhub are still installed in
container-beamsim-jupyter/container-conf/buid.sh. But, all scientific
packages (ex tensorflow) are now installed here.

This makes it easy to see everything in one place and make decisisions
to handle gpu vs cpu.

In addition, tensorflow was updated. We now just install tensorflow
and have pip handle the dependencies. The previous behavior of
installing dependencies ourselves was to get around a bug we now hope
is resolved.