Fix #697 add neutronics_material_maker for openmc (#698)
Bringing this in caused conflicts with bluesky and impactt.
Conflicting deps were pulled into common. While resolving the
conflicts I found bluesky could not be rebuilt. There were type errors
related to building scikit-beam and its use of cython. I decided to
remove scikit-beam. It isn't used in Sirepo and a cursory search among
users in our system revealed no uses. It was initially added for the
Raydata project but we now use a Conda environment from NSLS-II to
provide the deps.