Fix #703: vagrant-rsconf-dev working for almalinux (#709)
This installer is going to be called by rsconf to setup docker on rsconf
managed hosts. We only want to to do the tls, systemd, daemon.json, and
systemctl steps on "dev" hosts. On rsconf managed hosts rsconf will
manage the files and do the systemctl's.

Fix #705: centos7 is eol so need to use radiasoft depot mirror.
Fix #704: perl not yet installed so use sed
Fix #707: Flag to set dev vs rsconf environments.